requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
xhtml2pdf==0.2.15
reportlab==4.0.4
Pillow==10.4.0
python-dotenv==1.0.1